refactor: 도메인 관련 객체를 domain 모듈로 이동
