+++ match-api/build.gradle
@@ -1,7 +1,10 @@
 dependencies {
     implementation project(":match-core")
+    implementation project(":match-domain")
+    implementation project(":match-redis")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-webflux'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.8.9'
