+++ match-api/src/main/java/com/stocat/matchapi/MatchApiApplication.java
@@ -2,8 +2,10 @@ package com.stocat.matchapi;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = {"com.stocat.common", "com.stocat.matchapi"})
 public class MatchApiApplication {
 
     public static void main(String[] args) {
