+++ match-api/src/main/java/com/stocat/matchapi/websocket/OrderbookSubscriber.java
@@ -1,16 +1,13 @@
 package com.stocat.matchapi.websocket;
 
-import com.stocat.matchapi.domain.orderbook.Orderbook;
+import com.stocat.matchdomain.orderbook.Orderbook;
 import reactor.core.publisher.Flux;
 
 /**
  * Orderbook 구독자
- * - 외부 거래소 WebSocket 연결
- * - 실시간 호가 수신 및 파싱
- * - 종목별로 별도의 WebSocket 연결
+ * - 실시간 호가 데이터 스트림 제공
+ * - 구독 해제는 반환된 Flux의 Disposable로 관리 (Router 책임)
  */
 public interface OrderbookSubscriber {
     Flux<Orderbook> subscribe(String symbol);
-
-    void disconnect(String symbol);
 }
