+++ match-domain/build.gradle
@@ -1,0 +1,7 @@
+// Build as a plain library JAR
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    implementation project(":match-core")
+}
