+++ match-domain/src/main/java/com/stocat/matchdomain/AssetsCategory.java
@@ -1,4 +1,4 @@
-package com.stocat.matchapi.domain;
+package com.stocat.matchdomain;
 
 public enum AssetsCategory {
     KOR_STOCK, US_STOCK, CRYPTO, USD, KRW
