+++ match-domain/src/main/java/com/stocat/matchdomain/TradeSide.java
@@ -1,4 +1,4 @@
-package com.stocat.matchapi.domain;
+package com.stocat.matchdomain;
 
 /**
  * 매매 구분
