+++ match-domain/src/main/java/com/stocat/matchdomain/fill/Fill.java
@@ -1,6 +1,6 @@
-package com.stocat.matchapi.domain.fill;
+package com.stocat.matchdomain.fill;
 
-import com.stocat.matchapi.domain.TradeSide;
+import com.stocat.matchdomain.TradeSide;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
