+++ match-redis/build.gradle
@@ -1,4 +1,6 @@
 dependencies {
+	implementation project(":match-domain")
+
 	implementation 'org.springframework.boot:spring-boot-starter-webflux'
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
 }
