+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'stocat-match'
 
 include 'match-api'
 
+include 'match-domain'
 include 'match-redis'
 include 'match-mysql'
 include 'match-core'
